Ex 4 10: Reimplement RPN using getline
Exercise 4 10 asks me to reorganize the RPN calculator to use getline
rather than the getch/ungetch buffer. I implemented this using external
variables to maintain the current line and index in the line across
calls to getop. I actually consider this inferior to the getch/ungetch
solution because it couples getop strongly to the method for handling
input.