Ex 4 12: Implement recursive itoa function
Ex 4 12 asks me to "write a recursive version of itoa, that is, convert
an integer into a string by calling a recursive routine". I initially
tried to write this with a single function but ran into the problem of
completing the work after exiting recursion, namely null-terminating the
string and resetting the external index variable. I wrote an auxilary
recursive function to solve this, but I'm still wondering if there is a
simple single function solution.