Ex 4 13: Implement recursive string reversal
Ex 4 13 asks me to "write a recursive version of the function
reverse(s), which reverses the string s in place". Again I used an
auxiliary function for this, in this case to handle the setup work on
the index variables. This time I avoided using globals to pass
information down the recursion. I initially had a segmentation fault
when I incremented i (decremented j) instead of passing i+1 (j-1)... an
obvious mistake in retrospect. Yay learning!