Ex 5 01: Push signed nondigits back to input
Ex 5 01 asks me to amend the 'getint' function so that '-' or '+'
followed by a nondigit is no longer treated as a valid representation of
zero, and is instead pushed back onto the input. The effect is that the
appearance of an input like '-a' will now halt the scan, filling the
rest of the target array with zeroes. I don't really understand why this
is desirable behavior though.