Ex 5 04: Implement strend function
Ex 5 04 asks me to 'write the function strend(s, t), which returns 1 if
t occurs at the end of s, and zero otherwise'. I found this difficult,
and the code includes a commented out version I tried initially. That
method fails when the pattern (or actually any prefix of the pattern,
including the first letter) occurs more than once in the target.

My final version uses strlen from string.h, but it would be easy to
instead scan the auxiliary pointers to the end of the strings and proceed
the same way.