Ex 5 07: Rewrite readlines using array storage
Ex 5 07 asks me to 'rewrite readlines to store lines in an array
supplied by main rather than calling alloc to maintain storage' and then
time both versions. Rewriting readlines was not so difficult, though
this was the first time I had to use the & operator. I'm not sure how to
do a good time of timing the functions yet. I can cat a file to the
program and get the time for sorting that file once, but I'm not sure
how to sort the same file multiple times. This will probably be clearer
after reading chapter 8.