Ex 5 08: Error checking for calendar functions
Ex 5 08 asks me to add error checking to the functions 'day_of_year' and
'month_day' developed in section 5-7. I added checks to ensure that the
month is in the range 1 <= m <= 12 and the day is valid for the chosen
month. For 'month_day' I check that the desired day of the year is
in the range 1 <= d <= 365(366) for normal(leap) years.