Ex 5 10: Write command-line RPN calculator
Ex 5 10 asks me to 'write the program expr, which evaluates a reverse
Polish expression from the command line'. It was not so hard to alter
main and getop from the earlier exercise (ex 4 07) to use command line
arguments. Annoyingly, the multiply command * overlaps with the bash
shell wildcard, so it needs to be escaped. The simplest solution is
probably just to change the commands to avoid overlapping with bash
special characters (e.g. * -> x). Since special characters are parsed
before being passed to the program, I don't think there is any way to
fix this in my code, but I'd love to discover otherwise.