Clean up ex_1_21.c
I'm working on Ex 5 11, which entails revisiting the entab and detab
programs from ex 1 20 and ex 1 21. God help me, I've grown accustomed to
the terse style of K&R C, and feel compelled to make this code
compliant.

In the process I find that there is a bug in this code. It looks like
entab is inserting tabs inappropriately in some cases. I found this by
running the program on its source code. I'll fix this before proceeding
with chapter 5.