Ex 5 14: Add reverse option to sortlines
Ex 5 14 asks me to 'modify the sort program to handle a -r flag, which
indicates sorting in reverse (decreasing) order.' Initially I thought
the most straightforward thing was to reverse the array of strings after
sorting. However, I ran into trouble with the pointer arithmetic, and
saw strange things like truncated lines. I believe that the issue is
that when using alloc, the sizes of the elements are not predictable so
the pointer arithmetic is not accurate.

I then abandoned this approach and instead rewrote qsort slightly to
reverse the comparison operator. This should work for any comparison
function that conforms to the standard of strmp.