Ex 5 15: Add folding option to sortlines
Ex 5 15 asks me to 'add the option -f to fold upper and lower case
together, so that case distinctions are not made during sorting'. My
initial attempt was to convert all the strings to lower case before
sorting, but I decided that this was not really the desired behavior -
sorting shouldn't mutate the strings. Instead I wrote a new comparison
function that makes lowercase copies of the strings and compares them.