Ex 5 16: Add directory order option to sortlines
Ex 5 16 asks me to 'add the directory order option, which makes
comparisons only on letters, numbers, and blanks'. I again implemented
this with additional comparison operators. I'm using nested conditional
ternaries to handle the selection of the comparison, which is getting
messy, and I had to write some short 'lambda-like' functions to pass
options into the comparison so that I could use both -d and -f options
at the same time. This method is not scalable to more options.