Ex 6 2: Implement groupwords
Ex 6 2 asks me to 'write a program that reads a C program and prints in
alphabetical order each group of variable names that are identical in
the first 6 characters, but different somewhere thereafter' and 'make 6
a parameter that can be set from the command line.' This took a while.

The main difficulty turned out to be writing a function to convert a
binary tree into a sorted list. First, I was conflating a[][] with **a
in the type signatures, with attendant errors. Then I had difficulty
with correctly incrementing the pointers across the recursive function
calls. I elected to pass along an index variable to resolve this, which
was useful anyway for returning the number of words at the end.