Ex 6 3: Implement concordance maker
Ex 6 3 asks me to 'write a cross-referencer that prints a list of all
words in a document, and, for each word, a list of the line numbers on
which it occurs. Remove noise words like "the," "and," and so on.'

I've implemented the word to line concordance. I got confused about the
pointers again but worked through it rigorously and I believe it is now
correct. The main thing was recalling that e.g. int (*lines)[MAXLINES]
is correct for a pointer to an array of ints, while int *lines[MAXLINES]
gives an array of int pointers. Then I stripped out the c code specific
aspects of getword since this is for general documents. I still need to
filter out the noise words though.